Make common DB
